Added the time "Period" to the timestamp of each log line and to the opening status.  Unlike the startup/cycle number, the period only increments when the batteries were removed, which means that we don't know how much time has passed.  Within any one period, the poweredDays and RTC should reflect actual time that passed; between periods. This helps us track usage patterns.
The created/modified year for a file is the period + 2000, which is 
what we used to do for cycle, but period changes less often and might
be more helpful in this spot.  Note that the month and monthday were
set by an earlier commit to reflect real time (although time is frozen
when batteries are removed). So, month/monthday must be in synch with
poweredDays.  A future commit will likely force this to stay true.

Period is not incremented for a cold_start that occurred from a system
reset, because we now check for that flag (the log will no longer 
report battery removal if the reset flag was set).

Created #defines: FILE_YEAR_MIN 2000 and 
                  CLOCK_PERIOD (systemCounts.year - FILE_YEAR_MIN + 1)

Finally, added a log message when firmware img is found and when 
halting after cold start.