Reformat by trigger file, 'reformat.cmd'.
If the file exists in the root directory, the firmware will reformat the SD card, then
go into device mode, so that the Talking Book can be re-installed with a deployment.

After reformat, set volume label from SRN.

Read the SRN from the NOR flash, and if it starts with 'A-' or 'B-', set that as the
volume label. Also create the file a:/system/SERIALNO.srn on the SD card, so that the
TB-Loader will be able to identify the serial number, even with no other stats or files.

The volume label is written to the FAT32 boot blocks, and to the root directory. See
FAT32 docs online for more information.